new usage example fill rectangle
