--- public/usage-examples/graphics/draw_triangle/draw_triangle-1-random-triangles.txt
@@ -1,3 +1,0 @@
-#### Draw Triangle
-
-The following code shows examples of using [Draw Triangle](/api/graphics/#draw-triangle-1) to show how to draw random triangles.
