+++ public/usage-examples/graphics/fill_rectangle/fill_rectangle-1-simple.cpp
@@ -2,10 +2,14 @@
 
 int main()
 {
-    open_window("Fill rectangle", 800, 600);
-
+    //Create a window with title Fill rectangle, width 800, height 600.
+    open_window("Fill Rectangle", 800, 600);
     clear_screen();
-    fill_rectangle(COLOR_BLUE, 200, 200, 200, 100);
+
+    //draw three rectangles with color green, yellow, and red with different position as (100,200), (300,200), (500,200) with same size.
+    fill_rectangle(COLOR_GREEN, 100, 200, 200, 100);
+    fill_rectangle(COLOR_YELLOW, 300, 200, 200, 100);
+    fill_rectangle(COLOR_RED, 500, 200, 200, 100);
     refresh_screen();
     
     delay(4000);
