+++ public/usage-examples/graphics/fill_rectangle/fill_rectangle-1-simple.txt
@@ -1,3 +1,3 @@
-#### Basic Blue Rectangle
+#### Different Color Rectangles
 
-The following code shows examples of using [Fill Rectangle](/api/graphics#fill-rectangle-3) to show how to draw a filled-in blue rectangle.
+The following code shows examples of using [Fill Rectangle](/api/graphics/#fill-rectangle) to show how to draw a three different color rectangles side by side.
